% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi.R
\name{ndvi}
\alias{ndvi}
\title{Calculate NDVI}
\usage{
ndvi(NIR, red)
}
\arguments{
\item{NIR}{the value of near infrared band}

\item{red}{the value of red band}
}
\description{
Calculate NDVI
}
\examples{
ndvi(NIR = 25, red =50)
}
